consolidate ndjson.ts into common/xhr
