+++ ui/analyse/src/explorer/explorerXhr.ts
@@ -1,6 +1,6 @@
 import { ExplorerDb, OpeningData, TablebaseData } from './interfaces';
 import * as xhr from 'common/xhr';
-import { readNdJson } from 'common/ndjson';
+import { readNdJson } from 'common/xhr';
 import { ExplorerConfigData } from './explorerConfig';
 import { FEN } from 'chessground/types';
