+++ ui/ceval/src/engines/externalEngine.ts
@@ -1,6 +1,6 @@
 import { Work, ExternalEngineInfo, CevalEngine, CevalState, EngineNotifier } from '../types';
 import { randomToken } from 'common/algo';
-import { readNdJson } from 'common/ndjson';
+import { readNdJson } from 'common/xhr';
 import { throttle } from 'common/timing';
 
 interface ExternalEngineOutput {
